+++ test/test_helper.rb
@@ -1,10 +1,19 @@
-# Configure Rails Envinronment
 ENV["RAILS_ENV"] = "test"
+ENV["DB"] ||= "sqlite"
+
+unless File.exists?(File.expand_path('../../test/dummy/config/database.yml', __FILE__))
+  warn "WARNING: No database.yml detected for the dummy app, please run `rake prepare` first"
+end
+
+def using_mysql?
+  @using_mysql ||= ActiveRecord::Base.connection_config[:adapter].to_sym == :mysql2
+end
 
 require File.expand_path("../dummy/config/environment.rb",  __FILE__)
 require "rails/test_help"
 require 'shoulda'
 require 'ffaker'
+require 'database_cleaner' if using_mysql?
 
 Rails.backtrace_cleaner.remove_silencers!
 
@@ -14,9 +23,18 @@ ActiveRecord::Migrator.migrate File.expand_path("../dummy/db/migrate/", __FILE__
 # Load support files
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }
 
+# DatabaseCleaner is apparently necessary for doing proper transactions within MySQL (ugh)
+DatabaseCleaner.strategy = :truncation if using_mysql?
+
 # global setup block resetting Thread.current
 class ActiveSupport::TestCase
+  if using_mysql?
+    self.use_transactional_fixtures = false
+    setup { DatabaseCleaner.start }
+  end
+
   teardown do
+    DatabaseCleaner.clean if using_mysql?
     Thread.current[:paper_trail] = nil
   end
 end
@@ -32,4 +50,20 @@ def change_schema
     add_column :versions, :custom_created_at, :datetime
   end
   ActiveRecord::Migration.verbose = true
+  reset_version_class_column_info!
+end
+
+def restore_schema
+  ActiveRecord::Migration.verbose = false
+  ActiveRecord::Schema.define do
+    add_column :widgets, :sacrificial_column, :string
+    remove_column :versions, :custom_created_at
+  end
+  ActiveRecord::Migration.verbose = true
+  reset_version_class_column_info!
+end
+
+def reset_version_class_column_info!
+  PaperTrail::Version.connection.schema_cache.clear!
+  PaperTrail::Version.reset_column_information
 end