+++ test/unit/version_test.rb
@@ -7,6 +7,12 @@ class PaperTrail::VersionTest < ActiveSupport::TestCase
     assert PaperTrail::Version.creates.present?
   end
 
+  teardown do
+    restore_schema
+    Animal.connection.schema_cache.clear!
+    Animal.reset_column_information
+  end
+
   context "PaperTrail::Version.creates" do
     should "return only create events" do
       PaperTrail::Version.creates.each do |version|
@@ -59,10 +65,10 @@ class PaperTrail::VersionTest < ActiveSupport::TestCase
     setup { 2.times { @animal.update_attributes(:name => Faker::Lorem.word) } }
 
     context "receiving a TimeStamp" do
-      should "return all versions that were created before the Timestamp; descendingly by order of the `PaperTrail.timestamp_field`" do
-        value = PaperTrail::Version.subsequent(1.hour.ago)
+      should "return all versions that were created before the Timestamp" do
+        value = PaperTrail::Version.subsequent(1.hour.ago, true)
         assert_equal value, @animal.versions.to_a
-        assert_not_nil value.to_sql.match(/ORDER BY versions.created_at ASC\z/)
+        assert_not_nil value.to_sql.match(/ORDER BY #{PaperTrail::Version.arel_table[:created_at].asc.to_sql}/)
       end
     end
 
@@ -70,10 +76,6 @@ class PaperTrail::VersionTest < ActiveSupport::TestCase
       should "grab the Timestamp from the version and use that as the value" do
         value = PaperTrail::Version.subsequent(@animal.versions.first)
         assert_equal value, @animal.versions.to_a.tap { |assoc| assoc.shift }
-        # This asssertion can't pass in Ruby18 because the `strftime` method doesn't accept the %6 (milliseconds) command
-        if RUBY_VERSION.to_f >= 1.9 and not defined?(JRUBY_VERSION)
-          assert_not_nil value.to_sql.match(/WHERE \(versions.created_at > '#{@animal.versions.first.send(PaperTrail.timestamp_field).strftime("%F %T.%6N")}'\)/)
-        end
       end
     end
   end
@@ -82,10 +84,10 @@ class PaperTrail::VersionTest < ActiveSupport::TestCase
     setup { 2.times { @animal.update_attributes(:name => Faker::Lorem.word) } }
 
     context "receiving a TimeStamp" do
-      should "return all versions that were created before the Timestamp; descendingly by order of the `PaperTrail.timestamp_field`" do
-        value = PaperTrail::Version.preceding(Time.now)
+      should "return all versions that were created before the Timestamp" do
+        value = PaperTrail::Version.preceding(5.seconds.from_now, true)
         assert_equal value, @animal.versions.reverse
-        assert_not_nil value.to_sql.match(/ORDER BY versions.created_at DESC\z/)
+        assert_not_nil value.to_sql.match(/ORDER BY #{PaperTrail::Version.arel_table[:created_at].desc.to_sql}/)
       end
     end
 
@@ -93,10 +95,6 @@ class PaperTrail::VersionTest < ActiveSupport::TestCase
       should "grab the Timestamp from the version and use that as the value" do
         value = PaperTrail::Version.preceding(@animal.versions.last)
         assert_equal value, @animal.versions.to_a.tap { |assoc| assoc.pop }.reverse
-        # This asssertion can't pass in Ruby18 because the `strftime` method doesn't accept the %6 (milliseconds) command
-        if RUBY_VERSION.to_f >= 1.9 and not defined?(JRUBY_VERSION)
-          assert_not_nil value.to_sql.match(/WHERE \(versions.created_at < '#{@animal.versions.last.send(PaperTrail.timestamp_field).strftime("%F %T.%6N")}'\)/)
-        end
       end
     end
   end